HSQLAdapter

+++ anyline-data-jdbc-dialect/anyline-data-jdbc-hsqldb/src/main/java/org/anyline/data/jdbc/hsqldb/HSQLAdapter.java
@@ -48,7 +48,7 @@ public class HSQLAdapter extends SQLAdapter implements JDBCAdapter, Initializing
 		} 
 		PageNavi navi = run.getPageNavi(); 
 		if(null != navi){
-			int limit = navi.getLastRow() - navi.getFirstRow() + 1; 
+			long limit = navi.getLastRow() - navi.getFirstRow() + 1; 
 			if(limit < 0){
 				limit = 0; 
 			}
